Fix #41

+++ src/main/java/fi/felixbade/TelegramBotClient/APIModel/TelegramMessage.java
@@ -1,7 +1,7 @@
 package fi.felixbade.TelegramBotClient.APIModel;
 
 public class TelegramMessage {
-    public int message_id;
+    public long message_id;
     public TelegramUser from;
     public int date;
     public TelegramChat chat;
