+++ src/main/java/fi/felixbade/TelegramBotClient/TelegramBotClient.java
@@ -15,22 +15,22 @@ import fi.felixbade.TelegramBotClient.APIModel.TelegramUpdate;
 public class TelegramBotClient {
 
     public String token;
-    private int lastUpdateId = 0;
+    private long lastUpdateId = 0;
     private Gson gson = new Gson();
 
     public TelegramBotClient(String token) {
         this.token = token;
     }
 
-    public void sendMarkdownMessage(int telegramChatId, String messageToTelegram) {
+    public void sendMarkdownMessage(long telegramChatId, String messageToTelegram) {
         sendMessageWithParseMode(telegramChatId, messageToTelegram, "Markdown");
     }
 
-    public void sendMessage(int telegramChatId, String messageToTelegram) {
+    public void sendMessage(long telegramChatId, String messageToTelegram) {
         sendMessageWithParseMode(telegramChatId, messageToTelegram, "");
     }
 
-    public void sendMessageWithParseMode(int telegramChatId, String messageToTelegram, String parseMode) {
+    public void sendMessageWithParseMode(long telegramChatId, String messageToTelegram, String parseMode) {
         new Thread(new Runnable() {
             public void run() {
                 blockingSendMessage(telegramChatId, messageToTelegram, parseMode);
@@ -38,7 +38,7 @@ public class TelegramBotClient {
         }).start();
     }
 
-    public void blockingSendMessage(int telegramChatId, String messageToTelegram, String parseMode) {
+    public void blockingSendMessage(long telegramChatId, String messageToTelegram, String parseMode) {
         String url = String.format("https://api.telegram.org/bot%s/sendMessage", this.token);
 
         JsonObject blob = new JsonObject();
@@ -77,7 +77,7 @@ public class TelegramBotClient {
 
             for (JsonElement blob : updatesBlob) {
                 TelegramUpdate update = gson.fromJson(blob, TelegramUpdate.class);
-                int updateId = update.update_id;
+                long updateId = update.update_id;
                 if (updateId == this.lastUpdateId) {
                     continue;
                 }
